+++ packages/iso-did/src/common.js
@@ -6,7 +6,6 @@ export const KEY_TYPE_ALG = /** @type {const} */ ({
   'P-521': 'ES512',
   secp256k1: 'ES256K',
   RSA: 'RS256',
-  RSA_OLD: 'RS256_OLD',
 })
 
 export const ALG_KEY_TYPE = /** @type {const} */ ({
@@ -16,18 +15,16 @@ export const ALG_KEY_TYPE = /** @type {const} */ ({
   ES512: 'P-521',
   ES256K: 'secp256k1',
   RS256: 'RSA',
-  RS256_OLD: 'RSA_OLD',
 })
 
 /**
- * @typedef {keyof ALG_KEY_TYPE} Alg
+ * @typedef {keyof ALG_KEY_TYPE} SignatureAlgorithm
  * @typedef {keyof KEY_TYPE_ALG} KeyType
  */
 
 export const KEY_TYPE_CODE = /** @type {const} */ ({
   Ed25519: 0xed,
   RSA: 0x1205,
-  RSA_OLD: 11111,
   'P-256': 0x1200,
   'P-384': 0x1201,
   'P-521': 0x1202,
@@ -37,7 +34,6 @@ export const KEY_TYPE_CODE = /** @type {const} */ ({
 export const CODE_KEY_TYPE = /** @type {const} */ ({
   0xed: 'Ed25519',
   0x1205: 'RSA',
-  11111: 'RSA_OLD',
   0x1200: 'P-256',
   0x1201: 'P-384',
   0x1202: 'P-521',
@@ -56,7 +52,7 @@ export const CODE_KEY_TYPE = /** @type {const} */ ({
  */
 export function keyTypeToAlg(type) {
   // @ts-ignore
-  const alg = /** @type {Alg} */ (KEY_TYPE_ALG[type])
+  const alg = /** @type {SignatureAlgorithm} */ (KEY_TYPE_ALG[type])
 
   if (!alg) {
     throw new TypeError(`Unsupported key type ${type}`)
@@ -67,7 +63,7 @@ export function keyTypeToAlg(type) {
 /**
  * Signature algorithm name to key type.
  *
- * @param {Alg} alg
+ * @param {SignatureAlgorithm} alg
  * @throws {TypeError} if the algorithm is not supported.
  */
 export function algToKeyType(alg) {