+++ packages/iso-did/src/types.js
@@ -1,0 +1,2 @@
+/* eslint-disable unicorn/no-empty-file */
+// empty types