+++ packages/iso-signatures/src/utils.js
@@ -1,3 +1,6 @@
+import { DIDKey } from 'iso-did/key'
+import { equals, u8 } from 'iso-base/utils'
+
 /**
  * Create web crypto params for ECDSA.
  *
@@ -32,4 +35,31 @@ export function createEcdsaParams(alg) {
       throw new TypeError(`Unsupported algorithm ${alg}`)
     }
   }
+}
+
+/**
+ *
+ * @param {import('iso-did/types').KeyType} type
+ * @param {BufferSource} publicKey
+ * @param {import('iso-did/types').VerifiableDID} [did]
+ */
+export function didKeyOrVerifiableDID(type, publicKey, did) {
+  /** @type {import('iso-did/types').VerifiableDID} */
+  let _did = DIDKey.fromPublicKey(type, u8(publicKey))
+  if (did) {
+    if (!equals(did.publicKey, _did.publicKey)) {
+      throw new Error('Public key mismatch')
+    }
+
+    if (did.alg !== _did.alg) {
+      throw new Error('Algorithm mismatch')
+    }
+
+    if (did.type !== _did.type) {
+      throw new Error('Key type mismatch')
+    }
+    _did = did
+  }
+
+  return _did
 }