+++ packages/iso-did/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.5.0](https://github.com/hugomrdias/iso-repo/compare/iso-did-v1.4.0...iso-did-v1.5.0) (2023-12-21)
+
+
+### Features
+
+* verifiable did, document, resolver ([cce2203](https://github.com/hugomrdias/iso-repo/commit/cce220371903a93c366216cced7b215a0f0127ad))
+
+## [1.4.0](https://github.com/hugomrdias/iso-repo/compare/iso-did-v1.3.4...iso-did-v1.4.0) (2023-12-13)
+
+
+### Features
+
+* remove rsa old ([29ba18b](https://github.com/hugomrdias/iso-repo/commit/29ba18bdaca806c8c006697aa3e2b0572288c6e8))
+
+## [1.3.4](https://github.com/hugomrdias/iso-repo/compare/iso-did-v1.3.3...iso-did-v1.3.4) (2023-12-13)
+
+
+### Bug Fixes
+
+* expose types ([f66b567](https://github.com/hugomrdias/iso-repo/commit/f66b5670f621cc1b6f3a4a33e8231ecce92189a2))
+
 ## [1.3.3](https://github.com/hugomrdias/iso-repo/compare/iso-did-v1.3.2...iso-did-v1.3.3) (2023-09-05)