+++ packages/iso-signatures/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.3.0](https://github.com/hugomrdias/iso-repo/compare/iso-signatures-v0.2.0...iso-signatures-v0.3.0) (2023-12-21)
+
+
+### Features
+
+* update to new did types ([0db02d7](https://github.com/hugomrdias/iso-repo/commit/0db02d73e76e7c11cf6e60187482b2396e7ee244))
+
+## [0.2.0](https://github.com/hugomrdias/iso-repo/compare/iso-signatures-v0.1.10...iso-signatures-v0.2.0) (2023-12-13)
+
+
+### Features
+
+* expose types and export with wildcard ([de94912](https://github.com/hugomrdias/iso-repo/commit/de94912ba013dcff10ce33d7a0fd58b5a2161937))
+* remove rsa old ([29ba18b](https://github.com/hugomrdias/iso-repo/commit/29ba18bdaca806c8c006697aa3e2b0572288c6e8))
+
 ## [0.1.10](https://github.com/hugomrdias/iso-repo/compare/iso-signatures-v0.1.9...iso-signatures-v0.1.10) (2023-09-05)