+++ packages/iso-web/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.3.1](https://github.com/hugomrdias/iso-repo/compare/iso-web-v0.3.0...iso-web-v0.3.1) (2023-12-21)
+
+
+### Bug Fixes
+
+* export types ([93aa0a4](https://github.com/hugomrdias/iso-repo/commit/93aa0a409493bb01fc15b981a19822b416785dc6))
+
+## [0.3.0](https://github.com/hugomrdias/iso-repo/compare/iso-web-v0.2.0...iso-web-v0.3.0) (2023-12-15)
+
+
+### Features
+
+* updare msw to 2.0 ([fa934ae](https://github.com/hugomrdias/iso-repo/commit/fa934ae9d4d1f81dc74f15ada04dbe3621b38be4))
+
 ## [0.2.0](https://github.com/hugomrdias/iso-repo/compare/iso-web-v0.1.5...iso-web-v0.2.0) (2023-10-31)